# docs/swagger.yaml
definitions:
...
  dto.CreateUserDTO:
    properties:
      cnpj:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      number:
        type: string
      password:
        type: string
    required:
    - cnpj
    - email
    - first_name
    - last_name
    - number
    - password
    type: object
  dto.CreateUserResponse:
    properties:
      cnpj:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      number:
        type: string
    type: object
...
  dto.ReadEnterpriseResponse:
    properties:
      cnpj:
        type: string
      createdAt:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.ReadUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      enterprise:
        $ref: '#/definitions/dto.ReadEnterpriseResponse'
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      number:
        type: string
      updated_at:
        type: string
    type: object
  dto.ReadUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/dto.ReadUserResponse'
        type: array
    type: object
...
  dto.UpdateUserDTO:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      number:
        type: string
      password:
        type: string
    required:
    - email
    type: object
  dto.UpdateUserResponse:
    properties:
      updated:
        $ref: '#/definitions/dto.ReadUserResponse'
    type: object
  rest_err.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_err.RestErr:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_err.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Generic administrative API for a Basic Crud
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Basic CRUD API
  version: "1.0"
paths:
...
  /user/v1/:
    post:
      consumes:
      - application/json
      description: Create user by CNPJ, name and email
      parameters:
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - User
  /user/v1/{email}:
    delete:
      consumes:
      - application/json
      description: Delete user by email
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Read user by email
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ReadUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Read user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user by email (URI param) and update fields from body
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User
  /user/v1/read:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users
      parameters:
      - description: Page number (min 1)
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - User
  /user/v1/read/enterprise:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users
      parameters:
      - description: CNPJ enterprise
        in: query
        name: cnpj
        required: true
        type: string
      - description: Page number (min 1)
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: List users by CNPJ
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: 'Enter the token in the format: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"